src/factory_generator/requirements-dev.txt
jsonschema
PyYAML
pytest
coverage